pyproject.toml
[project]
name = "vectorgraph"
version = "0.1.0"
description = "Minimal Postgres stack combining Apache AGE graph + pgvector with simple async helpers."
authors = [
  { name = "QuixiAI" }
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
  "asyncpg>=0.29.0",
  "psycopg2-binary>=2.9.9",
  "python-dotenv>=1.0.0",
  "pgvector>=0.2.0",
  "numpy>=1.24.0",
  "tiktoken>=0.5.1",
  "openai>=1.3.0",
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0",
  "pydantic>=2.4.2",
]

[project.optional-dependencies]
dev = [
  "black>=23.11.0",
  "isort>=5.12.0",
  "mypy>=1.7.0",
  "pytest>=7.4.3",
  "pytest-asyncio>=0.21.1",
  "pytest-mock>=3.11.1",
  "pytest-cov>=4.1.0",
  "httpx>=0.25.0",
]

[project.scripts]
vectorgraph = "vectorgraph.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["vectorgraph"]
include-package-data = true

[tool.setuptools.package-data]
vectorgraph = ["stack/*", "examples/*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
